ofz#42501439 subtract the parent coordinate height in 64 bits
 MSPUBParser.cpp:1969:69: runtime error: signed integer overflow: 111102520 - -2038645869 cannot be represented in type 'int'
     #0 in libmspub::MSPUBParser::parseEscherShape(librevenge::RVNGInputStream*, libmspub::EscherContainerInfo const&, libmspub::Coordinate&, libmspub::Coordinate&) src/lib/MSPUBParser.cpp:1969
     #1 in libmspub::MSPUBParser::parseShapeGroup(librevenge::RVNGInputStream*, libmspub::EscherContainerInfo const&, libmspub::Coordinate, libmspub::Coordinate, unsigned int) src/lib/MSPUBParser.cpp:1606
     #2 in libmspub::MSPUBParser::parseShapeGroup(librevenge::RVNGInputStream*, libmspub::EscherContainerInfo const&, libmspub::Coordinate, libmspub::Coordinate, unsigned int) src/lib/MSPUBParser.cpp:1602
     #3 in libmspub::MSPUBParser::parseEscher(librevenge::RVNGInputStream*) src/lib/MSPUBParser.cpp:1581

the width beside it got the same in "ofz#6410 avoid signed integer overflow"

Reviewed-on: https://gerrit.libreoffice.org/c/libmspub/+/211232
Tested-by: Caolán McNamara <caolan.mcnamara@collabora.com>

+++ src/lib/MSPUBParser.cpp
@@ -1966,7 +1966,7 @@ void MSPUBParser::parseEscherShape(librevenge::RVNGInputStream *input, const Esc
             int coordSystemWidth = int64_t(thisParentCoordinateSystem.m_xe) - thisParentCoordinateSystem.m_xs;
             if (coordSystemWidth == 0)
               coordSystemWidth = 1;
-            int coordSystemHeight = thisParentCoordinateSystem.m_ye - thisParentCoordinateSystem.m_ys;
+            int coordSystemHeight = int64_t(thisParentCoordinateSystem.m_ye) - thisParentCoordinateSystem.m_ys;
             if (coordSystemHeight == 0)
               coordSystemHeight = 1;
             int groupWidth = int64_t(parentGroupAbsoluteCoord.m_xe) - parentGroupAbsoluteCoord.m_xs;
